requirements.txt
aiogram==3.7.0
telethon==1.36.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
aiohttp==3.9.5
python-dotenv==1.0.1
